requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
SQLAlchemy==2.0.30
asyncmy==0.2.9
pydantic-settings==2.3.4
python-dotenv==1.0.1
alembic==1.13.1
pytest==8.2.2
